MySql database was implemented and works fine

+++ src/main/java/com/example/projectmanagement/dto/ChartData.java
@@ -4,5 +4,5 @@ public interface ChartData {
 
     String getLabel();
 
-    long getNumber();
+    Integer getNumber();
 }
